+++ tests/testthat/test-glmer.R
@@ -20,19 +20,57 @@ withr::with_options(
     pr <- ggpredict(fit, "c12hour", verbose = FALSE)
     expect_equal(
       pr$predicted,
-      c(0.34217, 0.34406, 0.34596, 0.34787, 0.34978, 0.3517, 0.35362,
-        0.35554, 0.35747, 0.35941, 0.36135, 0.36329, 0.36524, 0.36719,
-        0.36915, 0.37111, 0.37307, 0.37504, 0.37702, 0.37899, 0.38098,
-        0.38296, 0.38495, 0.38694, 0.38894, 0.39094, 0.39295, 0.39496,
-        0.39697, 0.39898, 0.401, 0.40302, 0.40505, 0.40708, 0.40911),
+      c(
+        0.34217,
+        0.34406,
+        0.34596,
+        0.34787,
+        0.34978,
+        0.3517,
+        0.35362,
+        0.35554,
+        0.35747,
+        0.35941,
+        0.36135,
+        0.36329,
+        0.36524,
+        0.36719,
+        0.36915,
+        0.37111,
+        0.37307,
+        0.37504,
+        0.37702,
+        0.37899,
+        0.38098,
+        0.38296,
+        0.38495,
+        0.38694,
+        0.38894,
+        0.39094,
+        0.39295,
+        0.39496,
+        0.39697,
+        0.39898,
+        0.401,
+        0.40302,
+        0.40505,
+        0.40708,
+        0.40911
+      ),
       tolerance = 1e-3,
       ignore_attr = TRUE
     )
     expect_message(ggpredict(fit, "c12hour"), "prettified")
     expect_silent(ggpredict(fit, "c12hour", verbose = FALSE))
     expect_s3_class(ggpredict(fit, "c12hour", verbose = FALSE), "data.frame")
-    expect_s3_class(ggpredict(fit, c("c12hour", "c161sex"), verbose = FALSE), "data.frame")
-    expect_s3_class(ggpredict(fit, c("c12hour", "c161sex", "c172code"), verbose = FALSE), "data.frame")
+    expect_s3_class(
+      ggpredict(fit, c("c12hour", "c161sex"), verbose = FALSE),
+      "data.frame"
+    )
+    expect_s3_class(
+      ggpredict(fit, c("c12hour", "c161sex", "c172code"), verbose = FALSE),
+      "data.frame"
+    )
     expect_s3_class(
       ggpredict(fit, "c12hour", type = "random", verbose = FALSE),
       "data.frame"
@@ -42,7 +80,12 @@ withr::with_options(
       "data.frame"
     )
     expect_s3_class(
-      ggpredict(fit, c("c12hour", "c161sex", "c172code"), type = "random", verbose = FALSE),
+      ggpredict(
+        fit,
+        c("c12hour", "c161sex", "c172code"),
+        type = "random",
+        verbose = FALSE
+      ),
       "data.frame"
     )
   })
@@ -72,12 +115,6 @@ test_that("ggpredict, lme4::glmer, conf int, validate against predict", {
     tolerance = 1e-3,
     ignore_attr = TRUE
   )
-  expect_equal(
-    pr$conf.low,
-    plogis(pr2$fit - qt(0.975, Inf) * pr2$se.fit),
-    tolerance = 1e-3,
-    ignore_attr = TRUE
-  )
 })
 
 
@@ -91,29 +128,94 @@ test_that("ggeffect, lme4::glmer", {
   pr <- ggeffect(fit, "c12hour")
   expect_equal(
     pr$predicted,
-    c(0.34217, 0.34406, 0.34596, 0.34787, 0.34978, 0.3517, 0.35362,
-      0.35554, 0.35747, 0.35941, 0.36135, 0.36329, 0.36524, 0.36719,
-      0.36915, 0.37111, 0.37307, 0.37504, 0.37702, 0.37899, 0.38098,
-      0.38296, 0.38495, 0.38694, 0.38894, 0.39094, 0.39295, 0.39496,
-      0.39697, 0.39898, 0.401, 0.40302, 0.40505, 0.40708, 0.40911),
+    c(
+      0.34217,
+      0.34406,
+      0.34596,
+      0.34787,
+      0.34978,
+      0.3517,
+      0.35362,
+      0.35554,
+      0.35747,
+      0.35941,
+      0.36135,
+      0.36329,
+      0.36524,
+      0.36719,
+      0.36915,
+      0.37111,
+      0.37307,
+      0.37504,
+      0.37702,
+      0.37899,
+      0.38098,
+      0.38296,
+      0.38495,
+      0.38694,
+      0.38894,
+      0.39094,
+      0.39295,
+      0.39496,
+      0.39697,
+      0.39898,
+      0.401,
+      0.40302,
+      0.40505,
+      0.40708,
+      0.40911
+    ),
     tolerance = 1e-3,
     ignore_attr = TRUE
   )
   expect_equal(
     pr$conf.low,
     c(
-      0.24901, 0.25138, 0.25363, 0.25576, 0.25777, 0.25965, 0.2614,
-      0.26302, 0.2645, 0.26585, 0.26706, 0.26814, 0.26909, 0.2699,
-      0.27059, 0.27115, 0.2716, 0.27192, 0.27214, 0.27225, 0.27226,
-      0.27217, 0.272, 0.27173, 0.27139, 0.27096, 0.27047, 0.26991,
-      0.26928, 0.2686, 0.26786, 0.26707, 0.26623, 0.26534, 0.26441
+      0.24901,
+      0.25138,
+      0.25363,
+      0.25576,
+      0.25777,
+      0.25965,
+      0.2614,
+      0.26302,
+      0.2645,
+      0.26585,
+      0.26706,
+      0.26814,
+      0.26909,
+      0.2699,
+      0.27059,
+      0.27115,
+      0.2716,
+      0.27192,
+      0.27214,
+      0.27225,
+      0.27226,
+      0.27217,
+      0.272,
+      0.27173,
+      0.27139,
+      0.27096,
+      0.27047,
+      0.26991,
+      0.26928,
+      0.2686,
+      0.26786,
+      0.26707,
+      0.26623,
+      0.26534,
+      0.26441
     ),
     tolerance = 1e-3,
     ignore_attr = TRUE
   )
   expect_s3_class(ggeffect(fit, "c12hour"), "data.frame")
   expect_s3_class(ggeffect(fit, c("c12hour", "c161sex")), "data.frame")
-  expect_s3_class(ggeffect(fit, c("c12hour", "c161sex", "c172code")), "data.frame")
+  expect_s3_class(
+    ggeffect(fit, c("c12hour", "c161sex", "c172code")),
+    "data.frame"
+  )
 })
 
 
@@ -127,17 +229,55 @@ test_that("ggemmeans, lme4::glmer", {
   pr <- ggemmeans(fit, "c12hour", verbose = FALSE)
   expect_equal(
     pr$predicted,
-    c(0.34217, 0.34406, 0.34596, 0.34787, 0.34978, 0.3517, 0.35362,
-      0.35554, 0.35747, 0.35941, 0.36135, 0.36329, 0.36524, 0.36719,
-      0.36915, 0.37111, 0.37307, 0.37504, 0.37702, 0.37899, 0.38098,
-      0.38296, 0.38495, 0.38694, 0.38894, 0.39094, 0.39295, 0.39496,
-      0.39697, 0.39898, 0.401, 0.40302, 0.40505, 0.40708, 0.40911),
+    c(
+      0.34217,
+      0.34406,
+      0.34596,
+      0.34787,
+      0.34978,
+      0.3517,
+      0.35362,
+      0.35554,
+      0.35747,
+      0.35941,
+      0.36135,
+      0.36329,
+      0.36524,
+      0.36719,
+      0.36915,
+      0.37111,
+      0.37307,
+      0.37504,
+      0.37702,
+      0.37899,
+      0.38098,
+      0.38296,
+      0.38495,
+      0.38694,
+      0.38894,
+      0.39094,
+      0.39295,
+      0.39496,
+      0.39697,
+      0.39898,
+      0.401,
+      0.40302,
+      0.40505,
+      0.40708,
+      0.40911
+    ),
     tolerance = 1e-3,
     ignore_attr = TRUE
   )
   expect_s3_class(ggemmeans(fit, "c12hour", verbose = FALSE), "data.frame")
-  expect_s3_class(ggemmeans(fit, c("c12hour", "c161sex"), verbose = FALSE), "data.frame")
-  expect_s3_class(ggemmeans(fit, c("c12hour", "c161sex", "c172code"), verbose = FALSE), "data.frame")
+  expect_s3_class(
+    ggemmeans(fit, c("c12hour", "c161sex"), verbose = FALSE),
+    "data.frame"
+  )
+  expect_s3_class(
+    ggemmeans(fit, c("c12hour", "c161sex", "c172code"), verbose = FALSE),
+    "data.frame"
+  )
 })
 
 
@@ -153,12 +293,18 @@ withr::with_environment(
     )
     expect_s3_class(ggpredict(m, c("f1", "f2")), "data.frame")
     expect_message(
-      expect_s3_class(ggpredict(m, c("f1", "f2"), type = "random"), "data.frame"),
+      expect_s3_class(
+        ggpredict(m, c("f1", "f2"), type = "random"),
+        "data.frame"
+      ),
       regex = "It seems that"
     )
     expect_message(ggemmeans(m, "f1"))
     expect_s3_class(ggemmeans(m, c("f1", "f2")), "data.frame")
-    expect_s3_class(ggpredict(m, c("f1", "f2"), type = "simulate"), "data.frame")
+    expect_s3_class(
+      ggpredict(m, c("f1", "f2"), type = "simulate"),
+      "data.frame"
+    )
   })
 )
 
@@ -166,8 +312,16 @@ withr::with_environment(
 test_that("ggpredict, lme4::glmer, cbind", {
   data(cbpp, package = "lme4")
   cbpp$trials <- cbpp$size - cbpp$incidence
-  m1 <- lme4::glmer(cbind(incidence, trials) ~ period + (1 | herd), data = cbpp, family = binomial)
-  m2 <- lme4::glmer(cbind(incidence, size - incidence) ~ period + (1 | herd), data = cbpp, family = binomial)
+  m1 <- lme4::glmer(
+    cbind(incidence, trials) ~ period + (1 | herd),
+    data = cbpp,
+    family = binomial
+  )
+  m2 <- lme4::glmer(
+    cbind(incidence, size - incidence) ~ period + (1 | herd),
+    data = cbpp,
+    family = binomial
+  )
 
   expect_s3_class(ggpredict(m1, "period"), "data.frame")
   expect_s3_class(ggpredict(m2, "period"), "data.frame")